requirements.txt
pywin32==223
scikit-learn==0.20.3
python.dateutil==2.8.0
numpy==1.16.3
pandas==0.24.2
dill==0.2.9
simplegeneric==0.8.1
python.shutil
python.json
python.os
python.sys
python.glob
python.email
python.smtplib